Version 0.14 - New power up & needed updates
I removed the speed power up as it was annoying some people, and it is more of a power down.

The game will now show different high scores for each grid size and speed.

There is now a new powerup called score multiplier! It adds 2 to your score instead of one when food is eaten.

You now get a warning just before the powerups end.